Refactor: move processing methods to utils
